pyproject.toml
[tool.poetry]
name = "runpod-serverless-template"
version = "0.1.0"
description = "Template package for RunPod serverless endpoint deployment"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "runpod_serverless_template"}]

[tool.poetry.dependencies]
python = "^3.10"
runpod = "^1.3.0"
numpy = "^1.26.0"
pillow = "^10.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
python-dotenv = "^1.0.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
